--- src/components/HistoryPage/Button.jsx
@@ -1,10 +1,0 @@
-import React from 'react';
-import styles from '../../styles/HistoryStyles'; // Assuming the styles are in the HistoryStyles.js file
-
-const Button = ({ label, onClick }) => (
-  <button style={styles.button} onClick={onClick}>
-    {label}
-  </button>
-);
-
-export default Button;
