+++ src/components/card.jsx
@@ -1,4 +1,4 @@
-import React, { useState } from 'react';
+import { useState } from 'react';
 import '../styles/card.css';
 
 //create a functional component UserEntries
