--- src/styles/exerciseCard.css
@@ -1,40 +1,0 @@
-.exercise-card {
-  margin: 10px;
-  padding: 10px;
-  border: 1px solid #ddd;
-  border-radius: 8px;
-  /* background-color: #f9f9f9; */
-  box-shadow: 0px 2px 5px rgba(0, 0, 0, 0.1);
-  width: 18rem;
-  background-color: #7f8d96 !important;
-}
-
-.exercise-card h3 {
-  /* font-weight: bold; */
-  margin-top: 0;
-  color: #333;
-}
-
-.exercise-card button {
-  margin-right: 5px;
-}
-
-.delete-btn {
-  background-color: #48535b !important;
-  border-color: #48535b !important;
-}
-
-.delete-btn:hover {
-  background-color: #b4756e !important;
-  border-color: #b4756e !important;
-}
-
-.update-btn {
-  background-color: #b4756e !important;
-  border-color: #b4756e !important;
-}
-
-.update-btn:hover {
-  background-color: #48535b !important;
-  border-color: #48535b !important;
-}
